Refactoring network into swam, rate limiting and network types
